refactored and added javadoc #8 #30
