--- src/fileHandling/ILearningFile.tsx
@@ -1,9 +1,0 @@
-export interface ILearningFile{
-    fileName: string;
-    title: string;
-    zipCode: string;
-    city: string;
-    countryCode: string;
-    overpassAreaId: string;
-    startCoordinates: [number, number];
-}
