+++ src/learningFileHandling/LearningFile.ts
@@ -1,6 +1,8 @@
 import { ILearningFile } from "./ILearningFile";
-import { loadStreets } from "./FileHandler";
-
+import { fetchStreets } from "./FileHandler";
+/**
+ * Class representing a Learning File and it's corresponding street file
+ */
 export class LearningFile implements ILearningFile {
     fileName: string;
     title: string;
@@ -13,17 +15,24 @@ export class LearningFile implements ILearningFile {
     //arrays can't be null and I cant be sure nonempty so...
     private streetsLoaded: boolean = false;
 
-    
+    /**
+     * Load the streets File specified
+     * @returns The Promise from the Fetch containing the a List of SreetNames
+     */
     public async getStreets() : Promise<string[]> {
         if(this.streetsLoaded){
             return this.streets;
         }
         //change here once loader is programmed
-        this.streets = await loadStreets(this.fileName)
+        this.streets = await fetchStreets(this.fileName)
         this.streetsLoaded = true;
         return this.streets;
     }
     
+    /**
+     * Instanciate from ILearningFile Interface
+     * @param intrfce ILearning File to instanciate from
+     */
     constructor(intrfce:ILearningFile){
         //is this a proper way? 0.o
         this.fileName = intrfce.fileName;
